+++ src/Microsoft.ApplicationInsights/DataContracts/PageViewTelemetry.cs
@@ -5,6 +5,7 @@
     using System.Globalization;
     using System.Threading;
     using Microsoft.ApplicationInsights.Channel;
+    using Microsoft.ApplicationInsights.Extensibility;
     using Microsoft.ApplicationInsights.Extensibility.Implementation;
     using Microsoft.ApplicationInsights.Extensibility.Implementation.External;
 
@@ -21,10 +22,10 @@
     {
         internal const string TelemetryName = "PageView";
 
-        internal readonly string BaseType = typeof(PageViewData).Name;
-
+        internal readonly string BaseType = typeof(PageViewData).Name;        
         internal readonly PageViewData Data;
         private readonly TelemetryContext context;
+        private IExtension extension;
 
         private double? samplingPercentage;
 
@@ -55,6 +56,7 @@
         {
             this.Data = source.Data.DeepClone();
             this.context = source.context.DeepClone(this.Data.properties);
+            this.extension = source.extension?.DeepClone();
         }
 
         /// <summary>
@@ -73,6 +75,15 @@
         public TelemetryContext Context
         {
             get { return this.context; }
+        }
+
+        /// <summary>
+        /// Gets or sets gets the extension used to extend this telemetry instance using new strong typed object.
+        /// </summary>
+        public IExtension Extension
+        {
+            get { return this.extension; }
+            set { this.extension = value; }
         }
 
         /// <summary>